refactor: move Star component to core
