+++ app/core/components/Star.tsx
